+++ ci/test/test_ci.py
@@ -2,9 +2,6 @@ import asyncio
 import json
 import logging
 
-import pytest
-
-from ci.github import GithubStatus, github_status
 from hailtop.auth import hail_credentials
 from hailtop.config import get_deploy_config
 from hailtop.httpx import client_session
@@ -51,15 +48,3 @@ async def test_envoy_config_debug_endpoint():
             async with client_session() as session:
                 headers = await creds.auth_headers()
                 await retry_transient_errors(session.get_read_json, url, headers=headers)
-
-
-def test_github_check_statuses():
-    for state in ["PENDING", "EXPECTED", "ACTION_REQUIRED", "STALE"]:
-        assert github_status(state) == GithubStatus.PENDING
-    for state in ["FAILURE", "ERROR", "TIMED_OUT", "CANCELLED", "STARTUP_FAILURE", "SKIPPED"]:
-        assert github_status(state) == GithubStatus.FAILURE
-    for state in ["SUCCESS", "NEUTRAL"]:
-        assert github_status(state) == GithubStatus.SUCCESS
-    for state in ["foo", "bar"]:
-        with pytest.raises(ValueError):
-            github_status(state)
